enterprise_tools/requirements.txt
PySide6
qtawesome
PyPDF2
Pillow
pdf2image
PyMuPDF
pdf2docx
pywin32
psutil
python-docx
openpyxl
python-pptx
